models/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
google-generativeai
groq
pandas
numpy==1.24.3
faiss-cpu
sentence-transformers
pathlib
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
transformers